improve efficiency, convert points format from double 64 bit to float 32 bit

+++ script/generator.py
@@ -229,7 +229,7 @@ def generate_dataset(
     points = np.vstack(clusters)
     noise = gen_noise(n_noise, points, rng)
 
-    return np.vstack([points, noise])
+    return np.vstack([points, noise]).astype(np.float32)
 
 
 def save_dataset(index, data, n, c, center_scale, std_scale, noise_ratio):
